Split Doc parser from Unison lexer
